+++ l10n/oc.js
@@ -1,0 +1,8 @@
+OC.L10N.register(
+    "integration_reddit",
+    {
+    "Bad credentials" : "Marrits identificants",
+    "Connected accounts" : "Comptes connectats",
+    "Connected as {user}" : "Connectat coma {user}"
+},
+"nplurals=2; plural=(n > 1);");