+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## 1.0.7 – 2023-02-22
+### Changed
+- add 26 compat
+- lazy load dashboard widget
+- use @nextcloud/vue 7.6.1
+
+## 1.0.5 – 2022-08-29
+### Changed
+- implement proper token refresh based on expiration date
+- use material icons everywhere
+- make the app ready for NC 25 style changes
+- bump js libs, asjut to new eslint config
+
 ## 1.0.2 – 2021-09-13
 ### Changed
 - bump js libs